refactor: removed unnecessary file
